@@ -103,7 +103,7 @@ func (s *Suite) Test_Failed_Marktteilnehmer_Validation() {
 	VerfiyFailedValidations(s, validate, invalidMarktteilnehmer)
 }
 
-//  Test_Successful_Marktteilnehmer_Validation verifies that a valid BO is validated without errors
+// Test_Successful_Marktteilnehmer_Validation verifies that a valid BO is validated without errors
 func (s *Suite) Test_Successful_Marktteilnehmer_Validation() {
 	validate := validator.New()
 	validMarktteilnehmers := []bo.BusinessObject{
